Convert playlist integration test to example
